Added new test cases for games lost and bad moves the AI has taken
To correct the bad moves I have increased the depth to 6. In some cases
we weren't able to see far enough ahead to predict a loss. We also seem
to be well under the 0.5s per move.
In addition I fixed the the shortcut that returns the current score if
either player has won. This was based on a score of +-10000 but you then
lose information because some board states will setup multiple wins
which should score higher. The FieldScorer now returns the full score
and a separate boolean to say if either player has won.